backend/requirements.txt
sanic==0.7.0
tipsi-tools==1.36.*

flake8==3.5.0
pytest-asyncio==0.9.*
pytest-xdist==1.23.*
